Removed the Double Cast

+++ src/DumpItemsMod/CrpgItemValueModel.cs
@@ -15,7 +15,7 @@ namespace Crpg.DumpItemsMod
             double a = 0.05;
             double b = 50;
             double c = 100;
-            return (a * Math.Pow((double)itemTierf, 2) + b * (double)itemTierf + c);
+            return (float)(a * Math.Pow(itemTierf, 2) + b * itemTierf + c);
         }
     }
 }
